Projector state overhaul
Overhauls projectors to use a field and active state. Projectors that do not form a field will now use a static model instead of a TER, and will have field=inactive as a property.

In addition, these states are also datagenerated as well, which means that theoretically, a clever designer could make upgradable base models for the projectors now. The dish stays the same, though.